SoundWire: Create bubble text
This creates information "bubbles" on clock and data for
every frame.

Clock: shows parity check result and raw control word value.

Data: Show opcode type, ping status, rd/wr address and data,
      and decoded NAK+ACK bits.

As the displayable text length is limited by the bubble
width (which depends on zoom) the text is slightly shortened
by:
- Always use hexadecimal. There is no good reason to want any
  of the values shown in decimal.
- Omit 0x prefix on hex
- Abbreviate some of the labels.
- Put less important information on the right.
- ACK and NAK are decoded to a status string as this can
  be displayed more compactly than the separate ack and
  nak status values.

The OK/FAIL/IGNORED status is shown in full because normally
this will be "OK", and if it isn't it probably justifies
being entirely obvious what the status is.

Change-Id: Ic4c0f8fa2aa2572591655a61ca01159346ebecb8
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>