SoundWire: Add option to suppress duplicate pings from table
The Soundwire host will send a PING frame any time it doesn't
have a read or write to send. This causes a lot of noise in the
table view.

Add a settings option to suppress reporting of pings that are
unchanged from the previous ping. This only affects the FrameV2
objects used for table display. The bubble display and file export
will show all pings.

Change-Id: I12166f2ffb96d8ab4815d1c1aa5eae064c3c1e39
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>