SoundWire: Verify static sync pattern is a real frame
Check that a static sync pattern is genuine by trying to
find a sequence of frames that have the correct static
and dynamic sequence fields.

Don't assume that we will find a full 16-frame sequence.
It's possible that a capture could be a short fragment
that doesn't contain a full sequence (for example, it
might be the last few frames before a clock stop).
This initial implementation is simplistic and only two
consecutive good frames are required to decide that it
is in-sync.

Change-Id: I3226105ac8ef7bec81f4bceb0ed21117c31d125f
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>