SoundWire: Shrink the amount of memory used by bit history
The history buffer will typically need to be able to hold a
bus reset followed by a 16-frame sync sequence (and possibly
a little more than this). At the largest frame size this is

  4096 + (16 * 256 * 16) = 69632 bits

Storing a full U64 for every sample number would take
~544kBytes. With the 1-byte data level entry this is 612kB.

As the history is traversed linearly each entry could store
the delta from the previous sample number. This is much more
compact because SWIRE_CLK would normally be >1MHz so at
500MS/s there would be ~250 samples between data bits. This
would fit into a byte but using a U16 instead means that the
data level can be included in the same entry. This reduces
the typical worst-case  memory from 612kB to ~136kB. Apart
from saving memory it should increase speed because of higher
cache hits, and there is less data to move if the vector grows.

Sample deltas could be more than will fit into the U16 if
there is a gap - for example the last few frames before a
clock-stop. On bit of the U16 is reserved as an "overflow"
bit - if set the delta value continues into the next
history buffer entry.

Change-Id: Ib2fed2a9eb80f73d7665e66921466286c26fca8f
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>