SoundWire: Include sync loss entries in export file
The FrameV2 object shown in the UI table view has an entry to indicate
when there was a loss of sync. Include this in the export files.

Unfortunately the Saleae API uses V1 frames for the export file and
doesn't provide a function to get at the V2 frames, so the V1 frame
has to be added to the results for the exporting code to read it.

Change-Id: Ia025f3ce8d516c8a8f280b29bd53914979208249
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>