SoundWire: Remove dummy table entry
The "dummy" entry at the start of a table was to attempt to force the UI
to display the columns in the order we want. This isn't perfect but is
better than leaving the UI to make its own decision - which can be rather
random.

We now have a frame shape entry and there will always be one of these
before any real data entries. So the actions of the dummy entry can be
merged into this.

Change-Id: I0e76cef4b6d912d7b54cacce1c104b3f20bc0b41
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>