SoundWire: Annotate bus resets
Rearrange the code in CBitstreamDecoder so that it pushes bus-reset events
to SoundWireAnalyzer.

Contiguous logic-1s are now counted only when bits are pulled from the capture
data, not when traversing history. This guarantees that the bus reset will
only be notified once (when that raw data is first traversed). This prevents
annotating the same section of trace multiple times, or adding an annotation
that is out-of-time-order (which isn't allowed by the Saleae UI).

SoundWireAnalyzer will create a bubble Frame and a table FrameV2 to log
the bus reset. For bubble Frame the mType field is now used to distinguish
a bus-reset Frame from a normal Frame.

For file export the exporting of SoundWire frames is separated into a new
function exportNormalFrame(). This prevents the code becoming deeply nested
when adding different behaviour based on Frame::mType.

Change-Id: I1ff8d3edd19e5541b33775018c73aed6c4ab387a
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>