SoundWire: Annotate bit value as new bits are read
Move the annotation of 0/1 bit values from the main analyzer loop
to CBitstreamDecoder::NextBitValue().

This simplifies the logic. The annotation must only be added once per
sample, but the main loop can go backwards in time when there is a
sync loss. The sync-finder loop is a separate loop.

In all cases the CBitstreamDecoder::NextBitValue() is the source of the
bit values, and is where the decision is made to take a bit from history
od the next bit from the channel data. So the existing logic already
distinguishes whether the bit is new.

The annotation code is moved to a public inline function in
SoundWireAnalyzer.h. This allows the compiler to inline the code into
CBitstreamDecoder::NextBitValue() and optimize it.

Change-Id: Icc24549ee8160d36d39933f64f54e39b31d960b0
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>