SoundWire: Use fixed-size sliding search window for static sync
Use a fixed 4096-bit sliding window to search for the sync sequence and
move on by 4096 bits after each failed search.

Starting from any random position in the bitstream we must search the
equivalent of approximately two frames to find a sync. As we cannot know
how the starting point aligns with any frames, we might be starting part
way through a frame. In the worst case we just missed the first bit of the
static sync and must skip the rest of that frame to find the static sync
in the next frame. (Strictly speaking we don't need to search two full
frames, only one frame plus the length of a sync word, but searching two
full frames makes the maths, and the code, simpler.)

If the sync is not found within that 2-frame-sized window we cannot move
on by the full 2 frames because a static sync word could begin towards the
end of that window. In the worst case only the final bit of the sync word
sequence falls off the end of the search window, so we still need to
consider most of the second frame-sized window to be able to find that
sync word.

The previous method was to have a search window of 2-frame-size at each
column count, then move on by the smallest possible frame (96 bits).
But this means that many bits are being revisited multiple times for the
larger columns counts. For example, the maximum size frame of 4096 bits
would search a window of 8192 bits but then move on by only 96 bits so
8096 of those bits are revisited (8192 - 96), when in fact at most 4095
need to be revisted to find a sync word that fell off the end of the
window by only 1 bit.

A fixed-size sliding window of 4096 plus one frame reduces the number of
revisited bits. For the largest frame the search is still over 8192 bits
(4096 + one frame). But it then moves on by 4096 so only 4096 of the bits
are revisited.

Change-Id: I21a717c26de8aa896b0778b54e731947880b302f
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>