SoundWire: Use the data from the sample before the clock edge
In the SoundWire spec there is a very narrow window around clock edges
for when the data line is allowed to change. Data is allowed to change
state within 4ns of the clock edge, which is 2 samples at 500MS/s. This
can lead to the next data edge collapsing into the sample containing
the clock edge of the previous data state, thus giving the wrong value
for the data line at that clock edge.
As the data line can start to change within 4ns of the clock edge there
is usually a larger window before the clock edge where the data line
is stable at the correct state. So take the data value from the sample
before the clock edge.

Change-Id: Ieb125524f87a4515e719830df0b5e2aedc97afc0
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>