SoundWire: CSyncFinder: Optimize by testing all columns widths in parallel
When searching for the static sync sequence, there's no need to limit to
only trying one column count. The sync match checks for a pattern of bits
in a 128-bit shift register, and the only difference is the match pattern
and mask.

This patch changes CStaticSyncMatcher::PushBit() so that as each bit is
pushed into the accumulator it will be tested against the match sequence
for all possible column counts (unless a fixed column count has been
configured in the analyzer settings).

Because of this the static sync search no longer requires walking over the
same bits multiple times. It can continue to walk forwards through the bit
stream until a match is found.

Change-Id: I4a63564c2d890a5004b17256933f531c487ad9f2
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>