SoundWire: Don't count SSP as a difference when suppressing PINGs
Suppress pings that only differ by SSP bit.

When streaming or when the bus is idle there can be a large number of
PINSs that have the SSP bit set but are otherwise identical. This can
create a lot of entries in the table output that make the table view
difficult to scroll through.

Change-Id: Id47837173ee7c065bc69a178ddb220e59ce49742
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>